Updates the license owner
Whether having a license is necessary at all,
and what the license should be, are good questions.
I find it better to have a license than to have
no license, as explicitly adding one removes any
potential (and unecessary) source of confusion.

I chose Apache 2.0 since it is liberal, and easy
to deal with.